LayoutTests/fast/canvas/html-in-canvas/drawElementImage-2arg-snapshot-count.html

<!DOCTYPE html><!-- webkit-test-runner [ CanvasDrawElementEnabled=true ] -->
<title>drawElementImage snapshot count tracking (TB1b)</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<canvas id="c" layoutsubtree>
  <div id="a" style="background:red; width:50px; height:50px;"></div>
  <div id="b" style="background:blue; width:50px; height:50px;"></div>
</canvas>
<script>
function nextPaint() {
    return new Promise(r => requestAnimationFrame(() => requestAnimationFrame(r)));
}

promise_test(async () => {
    const canvas = document.getElementById("c");
    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2,
        "two direct children -> two snapshots after paint");
}, "snapshot count reflects direct children after paint");

promise_test(async () => {
    // Backing-store-only resize (canvas.width attribute change) preserves the
    // existing snapshots: the children's recorded paint is layout-driven, not
    // backing-store-driven. Pinned by the wpt_internal corpus test
    // draw-element-image-returned-matrix.html "Draw transform should account for
    // canvas pixel scale", which mutates canvas.width and immediately calls
    // drawElementImage with no frame await.
    const canvas = document.getElementById("c");
    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2);

    canvas.width = canvas.width + 1;
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2,
        "backing-store resize preserves snapshots");

    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2,
        "still two snapshots after next frame");
}, "canvas backing-store resize preserves child snapshots");

promise_test(async () => {
    const canvas = document.getElementById("c");
    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2);

    const a = document.getElementById("a");
    canvas.removeChild(a);
    assert_equals(internals.canvasChildSnapshotCount(canvas), 1,
        "removing a child immediately purges its snapshot");
    canvas.appendChild(a); // restore for any later tests in this file
}, "removing a direct child purges its snapshot");
</script>
</body>
